requirements.txt
fastapi
uvicorn[standard]
openai
openai-whisper
python-dotenv
ffmpeg-python
transformers
torch
accelerate